o more changes of pulse number representation from unsigned int to long long o restore data latency sleep() call before returning from   KaAfcPrivate::_processXmitAverage()
git-svn-id: http://svn.eol.ucar.edu/svn/ka/trunk@138 3b647ff1-7a5c-455a-93f8-4e24338e3c8e

+++ src/kadrx/KaAfc.h
@@ -46,7 +46,7 @@ public:
     /// @param g0Mag relative g0 power magnitude, in range [0.0,1.0]
     /// @param freqOffset measured frequency offset, in Hz
     /// @param pulsenum pulse number, counted since transmitter startup
-    void newXmitSample(double g0Mag, double freqOffset, unsigned int pulsenum);
+    void newXmitSample(double g0Mag, double freqOffset, long long pulsenum);
 protected:
     /// constructor for the singleton instance
     KaAfc();
